requirements._for_linux.txt
annotated-types==0.7.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
confection==0.1.5
cymem==2.0.11
faiss-cpu==1.12.0
filelock==3.19.1
fsspec==2025.9.0
greenlet==3.2.3
grpcio==1.75.1
grpcio-health-checking==1.75.1
grpcio-tools==1.75.1
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
langdetect==1.0.7
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nltk==3.9.2
numpy==2.3.3
#nvidia-cublas-cu12==12.8.4.1
#nvidia-cuda-cupti-cu12==12.8.90
#nvidia-cuda-nvrtc-cu12==12.8.93
#nvidia-cuda-runtime-cu12==12.8.90
#nvidia-cudnn-cu12==9.10.2.21
#nvidia-cufft-cu12==11.3.3.83
#nvidia-cufile-cu12==1.13.1.3
#nvidia-curand-cu12==10.3.9.90
#nvidia-cusolver-cu12==11.7.3.90
#nvidia-cusparse-cu12==12.5.8.93
#nvidia-cusparselt-cu12==0.7.1
#nvidia-nccl-cu12==2.27.3
#nvidia-nvjitlink-cu12==12.8.93
#nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==11.3.0
playwright==1.54.0
pluggy==1.6.0
preshed==3.0.10
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
PyYAML==6.0.3
rake-nltk==1.0.6
regex==2025.9.18
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
spacy==3.8.7
spacy-langdetect==0.1.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
#torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
#triton==3.4.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.14.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
